fix: определить интерфейс ItemService существующий класс ItemService переименовать в ItemServiceImpl, реализующим интерфейс ItemService в классе-контроллере ItemController заменить класс на интерфейс
